Rename 'views' folder to 'ui'
