+++ coronainfo/app.py
@@ -19,6 +19,7 @@ import logging
 import sys
 
 import gi
+
 gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
 from gi.repository import Gio, Adw
@@ -27,7 +28,7 @@ from gi.repository import Gio, Adw
 from coronainfo import _logger  # Unused but is necessary to initialise the logger
 from coronainfo.enums import App
 from coronainfo.settings import AppSettings
-from coronainfo.views import MainWindow, AboutDialog
+from coronainfo.ui import MainWindow, AboutDialog
 from coronainfo.utils.ui_helpers import create_action, log_action_call
 
 
@@ -77,12 +78,8 @@ class CoronaInfoApp(Adw.Application):
         self.quit()
 
 
-def get_schema() -> Gio.Settings:
-    return CoronaInfoApp.get_schema()
-
-
-def get_settings() -> AppSettings:
-    return CoronaInfoApp.get_settings()
+get_schema = CoronaInfoApp.get_schema
+get_settings = CoronaInfoApp.get_settings
 
 
 def main():
