+++ coronainfo/ui/dialog_about.py
