+++ coronainfo/ui/dialog_preferences.py
