+++ data/ui/help_overlay.ui
@@ -24,7 +24,7 @@
             <child>
               <object class="GtkShortcutsShortcut">
                 <property name="title" translatable="yes" context="shortcut window">Show shortcuts</property>
-                <property name="action-name">win.show-help-overlay</property>
+                <property name="action-name">win.show-help_overlay</property>
               </object>
             </child>
             <child>
