+++ data/ui/window_main.ui
@@ -122,7 +122,7 @@
       </item>
       <item>
         <attribute name="label" translatable="yes">_Keyboard shortcuts</attribute>
-        <attribute name="action">win.show-help-overlay</attribute>
+        <attribute name="action">win.show-help_overlay</attribute>
       </item>
       <item>
         <attribute name="label" translatable="yes">_About Corona Info</attribute>
